+++ packages/cluster-service/test/service.js
@@ -3,4 +3,4 @@ const http = require('http');
 http.createServer((req, res) => {
   res.writeHead(200, { 'Content-Type': 'text/plain' });
   res.end('test');
-}).listen(9615);
+}).listen(0);