+++ packages/gql-bot/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/restorecommerce/libs/compare/@restorecommerce/gql-bot@0.2.6...@restorecommerce/gql-bot@0.2.7) (2022-08-10)
+
+
+### Bug Fixes
+
+* **gql-bot:** always pre-process response before logging ([ff6963e](https://github.com/restorecommerce/libs/commit/ff6963e0f2c7c5c0d61a3d1f890706f52fe02d91))
+
+
+
+
+
+## [0.2.6](https://github.com/restorecommerce/libs/compare/@restorecommerce/gql-bot@0.2.5...@restorecommerce/gql-bot@0.2.6) (2022-08-10)
+
+
+### Bug Fixes
+
+* **gql-bot:** fix null type error ([22db1dc](https://github.com/restorecommerce/libs/commit/22db1dce1d3e31658852cb80194aa2140ecc0970))
+
+
+
+
+
+## [0.2.5](https://github.com/restorecommerce/libs/compare/@restorecommerce/gql-bot@0.2.4...@restorecommerce/gql-bot@0.2.5) (2022-08-10)
+
+**Note:** Version bump only for package @restorecommerce/gql-bot
+
+
+
+
+
 ## [0.2.4](https://github.com/restorecommerce/libs/compare/@restorecommerce/gql-bot@0.2.3...@restorecommerce/gql-bot@0.2.4) (2022-06-20)
 
 **Note:** Version bump only for package @restorecommerce/gql-bot