+++ packages/kafka-client/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.4](https://github.com/restorecommerce/kafka-client/compare/@restorecommerce/kafka-client@0.4.3...@restorecommerce/kafka-client@0.4.4) (2022-08-10)
+
+**Note:** Version bump only for package @restorecommerce/kafka-client
+
+
+
+
+
 ## [0.4.3](https://github.com/restorecommerce/kafka-client/compare/@restorecommerce/kafka-client@0.4.2...@restorecommerce/kafka-client@0.4.3) (2022-07-07)
 
 **Note:** Version bump only for package @restorecommerce/kafka-client