+++ packages/rc-grpc-clients/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.7](https://github.com/restorecommerce/libs/compare/@restorecommerce/rc-grpc-clients@0.5.6...@restorecommerce/rc-grpc-clients@0.5.7) (2022-08-10)
+
+**Note:** Version bump only for package @restorecommerce/rc-grpc-clients
+
+
+
+
+
 ## [0.5.6](https://github.com/restorecommerce/libs/compare/@restorecommerce/rc-grpc-clients@0.5.5...@restorecommerce/rc-grpc-clients@0.5.6) (2022-07-07)
 
 **Note:** Version bump only for package @restorecommerce/rc-grpc-clients